move DP related code
