+++ ipa-core/src/protocol/ipa_prf/oprf_padding/insecure.rs
@@ -5,7 +5,7 @@ use std::f64::consts::E;
 use rand::distributions::{BernoulliError, Distribution};
 use rand_core::{CryptoRng, RngCore};
 
-use crate::protocol::dp::distributions::{BoxMuller, RoundedBoxMuller, TruncatedDoubleGeometric};
+use crate::protocol::ipa_prf::oprf_padding::distributions::{BoxMuller, RoundedBoxMuller, TruncatedDoubleGeometric};
 
 #[derive(Debug, PartialEq, thiserror::Error)]
 pub enum Error {
