+++ ipa-core/src/protocol/ipa_prf/oprf_padding/mod.rs
@@ -1,0 +1,5 @@
+mod distributions;
+mod insecure;
+
+#[cfg(any(test, feature = "test-fixture", feature = "cli"))]
+pub use insecure::DiscreteDp as InsecureDiscreteDp;
